Meer bugfixes

+++ Code/Teacher/Assignments/deleteQuestion.php
@@ -2,16 +2,24 @@
 require_once "../../DB_Connection.php";
 session_start();
 
+//Check if there's only 1 question, if so, return an error
+$sth8 = $pdo -> prepare("SELECT * FROM `questions` WHERE assignmentID = ?");
+$sth8 -> execute([$_SESSION['editingAssign']]);
+$row8 = $sth8 -> fetch();
+
+if ($sth8 -> rowCount() == 1) {
+    $_SESSION['error'] = "Je kunt niet je enige vraag verwijderen.";
+    header("Location: assignmentEditor.php?assign=" . $_SESSION['editingAssign'] . "&question=" . $_SESSION['editingQuestion']);
+    exit();
+}
+
+//Delete media
 $sth1 = $pdo -> prepare("SELECT `media` FROM questions WHERE assignmentID = ? AND `order` = ?");
 $sth1 -> execute([$_SESSION['editingAssign'], $_GET['deleteQuestion']]);
 $row1 = $sth1 -> fetch();
 if (isset($row1['media']) && $row1['media'] != "") {
     unlink($row1['media']);
-    $sth2 = $pdo->prepare("UPDATE `questions` SET `media` = ? WHERE id = ?");
-    $sth2->execute(["", $_SESSION['editingAssign']]);
 }
-echo $_SESSION['editingAssign'];
-echo $_GET['deleteQuestion'];
 
 //Delete all elements in question
 $sth2 = $pdo -> prepare("DELETE FROM multiplechoices WHERE assignmentID = ? AND `questionOrder` = ?");
@@ -42,13 +50,9 @@ $sth8 = $pdo -> prepare("SELECT * FROM `questions` WHERE assignmentID = ?");
 $sth8 -> execute([$_SESSION['editingAssign']]);
 $row8 = $sth8 -> fetch();
 
-if ($_GET['editingQuestion'] == 1) {
-    $question = 2;
-} else {
-    $question = $_SESSION['editingQuestion'];
-    if ($question > $sth8 -> rowCount()) {
-        $question = $sth8 -> rowCount();
-    }
+$question = $_SESSION['editingQuestion'];
+if ($question > $sth8 -> rowCount()) {
+    $question = $sth8 -> rowCount();
 }
 
 header("Location: assignmentEditor.php?assign=" . $_SESSION['editingAssign'] . "&question=" . $question);
